pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "project"
requires-python = ">=3.7,<4.0"
version = "1.0"
dependencies = [
    # Add runtime dependencies here
]
readme = "README.md"
keywords = ["neuroscience", "rdm", "clip", "cnn", "pytorch", "tensorflow"]

[project.optional-dependencies]
dev = [
    "tox",
    "ipykernel",
    "project[ds]",
    "project[lint]",
    "project[build]",
]
ds = [
    "numpy==1.26.4",
    "pandas==2.2.2",
    "torch==2.5.1",
    "torchvision",
    "torchaudio",
    "tensorflow",
    "transformers",
    "scikit-learn",
    "scipy==1.13.1",
    "tqdm==4.67.1",
    "matplotlib==3.10.0",
    "seaborn==0.13.2",
    "Pillow"
]
lint = [
    "mypy",
    "ruff",
]

build = [
    "build>=1.2.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
src = ["src"]
extend-exclude = [
    "conf.py",
]
target-version = "py311"
lint.select = ["ALL"]
lint.ignore = [
    "COM812",   # Conflicts with the formatter
    "ISC001",   # Conflicts with the formatter
    "PT001",    # https://github.com/astral-sh/ruff/issues/8796#issuecomment-1825907715
    "PT023",    # https://github.com/astral-sh/ruff/issues/8796#issuecomment-1825907715
    "D415",     # Docstring can end without ./!
    "D416",     # Docstring Return can end without :
    "T201", # print statements
    "N802", # need to use provided func names which are camelCase
    "ANN001", # skip type annotation
    "UP031", # remove string format to allow for regex
    "PD901", # be able to use 'df' name
    "ANN201", # ignore return types
    "E501", # ignore line too long
    "D103", # ignore docstrings
    "PLR2004", # ignore magic number
    "PTH118", # ignore Path module requirement
    "D104", # ignore docstring
    "D100", # ignore docstring
    "ANN202",
    "ANN204", 
    "D105",
    "N806",
    "D107", # ignore docstring for __init__
    "D102", # ignore docstring in public method
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = [
    "S101", # Use of `assert` detected
    "D103", # Missing docstring in public function
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`(https://github.com/astral-sh/ruff/issues/5434)
keep-runtime-typing = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
disallow_untyped_defs = false
warn_unused_ignores = true
ignore_missing_imports = true
exclude = [
    "project-\\d+", # Ignore temporary folder created by setuptools when building an sdist
    "venv.*/",
    "build/",
    "dist/",
]

